test-practice/requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cowsay==3.0
cryptography==3.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
filelock==3.0.12
idna==3.1
iniconfig==1.1.1
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipython-sql==0.4.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.1
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prettytable==0.7.2
prometheus-client==0.9.0
prompt-toolkit==3.0.14
py==1.10.0
pycparser==2.20
Pygments==2.7.4
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.2
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
style==1.1.6
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
update==0.0.1
urllib3==1.26.3
virtualenv==20.4.2
wcwidth==0.2.5
webencodings==0.5.1
